Eliminated all clang compilation warnings

+++ dumb/src/it/readokt.c
@@ -322,8 +322,8 @@ unsigned get_chunk_count(IFF_CHUNKED *mod, unsigned type)
 static DUMB_IT_SIGDATA *it_okt_load_sigdata(DUMBFILE *f)
 {
 	DUMB_IT_SIGDATA *sigdata;
-	unsigned n_channels;
-	unsigned i, j, k, l;
+    int n_channels;
+    int i, j, k, l;
 	IFF_CHUNKED *mod;
 	const IFF_CHUNK *chunk;
 
@@ -431,7 +431,7 @@ static DUMB_IT_SIGDATA *it_okt_load_sigdata(DUMBFILE *f)
 	}
 
 	chunk = get_chunk_by_type(mod, DUMB_ID('P','A','T','T'), 0);
-	if (!chunk || chunk->size < sigdata->n_orders) {
+    if (!chunk || chunk->size < (unsigned)sigdata->n_orders) {
 		_dumb_it_unload_sigdata(sigdata);
 		free_okt(mod);
 		return NULL;
